+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## Version 1.1.1
+
+- Allow Bakerman to when the --workdir contains a valid repo and clone isn't
+  required anymore.
+
 ## Version 1.1.0
 
 - Make Bakerman more generic by not only focussing on Dockerfiles.